fix: repair HTTP symbol references broken by branch refactors
Two stragglers from this branch's own HTTP refactors were left with
stale references and broke CI:

- MockHTTPResponse moved from datadog_checks.base.utils.http_testing to
  datadog_checks.dev.http; update the four test files still importing
  the old path (sonatype_nexus, nifi, traefik_mesh, nutanix).
- HTTPResponseProtocol was renamed to HTTPResponse in http_protocol;
  update the two production files still referencing the old name
  (nutanix at runtime, control_m under TYPE_CHECKING).
