+++ control_m/changelog.d/22676.fixed
@@ -1,0 +1,1 @@
+Use the renamed backend-agnostic ``HTTPResponse`` protocol type hint.
