+++ control_m/datadog_checks/control_m/client.py
@@ -12,7 +12,7 @@ if TYPE_CHECKING:
     from collections.abc import Callable
 
     from datadog_checks.base import AgentCheck
-    from datadog_checks.base.utils.http_protocol import HTTPResponseProtocol
+    from datadog_checks.base.utils.http_protocol import HTTPResponse
 
 
 class ControlMClient:
@@ -115,7 +115,7 @@ class ControlMClient:
         self._log.info("Refreshing Control-M API token")
         self.login()
 
-    def request(self, method: str, url: str, **kwargs: Any) -> HTTPResponseProtocol:
+    def request(self, method: str, url: str, **kwargs: Any) -> HTTPResponse:
         # Route the request through static token or session login, with 401 fallback.
         request_fn = getattr(self._http, method)
 
@@ -140,9 +140,7 @@ class ControlMClient:
         self._static_token_retry_after = time.monotonic() + self._static_token_retry_interval
         return self._session_request(request_fn, url, **kwargs)
 
-    def _session_request(
-        self, request_fn: Callable[..., HTTPResponseProtocol], url: str, **kwargs: Any
-    ) -> HTTPResponseProtocol:
+    def _session_request(self, request_fn: Callable[..., HTTPResponse], url: str, **kwargs: Any) -> HTTPResponse:
         # Used in session login mode only. Make a request to the API using the session token.
         self.ensure_token()
         extra = {"Authorization": f"Bearer {self._token}"}
