+++ nutanix/tests/test_retry.py
@@ -6,7 +6,7 @@
 import pytest
 
 from datadog_checks.base.utils.http_exceptions import HTTPStatusError
-from datadog_checks.base.utils.http_testing import MockHTTPResponse
+from datadog_checks.dev.http import MockHTTPResponse
 from datadog_checks.nutanix import NutanixCheck
 
 pytestmark = [pytest.mark.unit]
